+++ src/main/java/net/oijon/oling/datatypes/grammar/GlossList.java
@@ -1,4 +1,4 @@
-package net.oijon.oling.datatypes;
+package net.oijon.oling.datatypes.grammar;
 
 import java.util.ArrayList;
