+++ src/main/java/net/oijon/oling/datatypes/grammar/Grammar.java
@@ -1,4 +1,4 @@
-package net.oijon.oling.datatypes;
+package net.oijon.oling.datatypes.grammar;
 
 /**
  * Reserved for grammar
