+++ src/main/java/net/oijon/oling/datatypes/orthography/Guesser.java
@@ -1,4 +1,4 @@
-package net.oijon.oling.datatypes;
+package net.oijon.oling.datatypes.orthography;
 
 public class Guesser {
