+++ src/main/java/net/oijon/oling/datatypes/tags/Multitag.java
@@ -1,4 +1,4 @@
-package net.oijon.oling.datatypes;
+package net.oijon.oling.datatypes.tags;
 
 import java.util.ArrayList;
