+++ src/main/java/net/oijon/oling/datatypes/tags/MultitagUtils.java
@@ -1,4 +1,4 @@
-package net.oijon.oling.datatypes;
+package net.oijon.oling.datatypes.tags;
 
 public class MultitagUtils {
